Fix a bug in instr:transform_receive_case/1
When transforming receive clauses to case clauses,
we search for a `catchall' pattern and if we don't find one
then we add ours. But Concuerror doesn't check if a variable is
bound or unbound and falsely mistake them as a `catchall' rule.
For now, always add a `catchall' rule.
Many thanks to David Klaftenegger for pointing this bug.