Rename every instrumented module
Append their names with `INSTR_PREFIX'.
This way we can instrument libraries that Concuerror
uses or modules that are sticky.

To do so we must open the file containing the source code,
change the `-module()' declaration and save it to a temp
directory. This has to be done because predefined macros
such `MODULE' has to be redefined (and this is the only
way erlang allows us to do that).

We have to keep a list of all the instrumented modules
and check at runtime if some function calls need to
be renamed.

When a ticket is produced, we clean the stacktrace to
hide the fact that we had renamed all the modules :)