Strip `callback' attribute in instrumentation
Callbacks allow behaviour modules to specify
(using a formal language, instead of just in comments) the names
and types of all functions that callback modules should implement.
This information can then be processed by dialyzer to check
conformance of a callback module with what the behaviour module
expects and warn the user for errors that may exist.

Modules such as `supervisor' can not be instrumented
because of this callbacks (they use `types' and we
already strip them).