Add --wait-messages flag
Concuerror will wait for messages from
uninstrumented processes to arrive.
This slows down Concuerror (in some cases
using this option can make a big difference
in time) so I have disable it by default.

Still there are applications that crash
with or without using this flag so we
need to find a better way to wait for
this kind of messages.