Remove conditional dependency on receive patterns
If message A was delivered to process P before message B, then there should
be a race only if the patterns in the receive that gets A can also match B.
Unfortunately this is a 'conditional dependency' (we don't know the patterns
when A or B is delivered) and cannot be handled correctly when planning more
interleavings. So this feature is removed for the time being.