Even more dependencies for messages
Matching sends with receives is problematic. This commit removes yet another
refinement that existed in the dependency relation which prevented reversing the
delivery of a message with a timeout-enabled receive if the message was not the
one actually received. It turns out that if this relation is used to detect
initials in the wakeup tree (where receives have not been updated to reflect
their new status) we miss interleavings.

Sad...