References are now a fixed sequence per process.
In a way similar to persistent PIDs for child processes, references are now also
monitored and reused. This is a workaround to avoid having dirrerent refs
for timers which are started from the 'same' state.

The deeper problem is that the dependency relation marks cannot (and should not)
tell apart calls to start_timer which are issued by the same state, but if the
original reference has been lost a new timer-simulating process will be
referenced as the one handling the timer. Apart from process pollution, this
also leads to the new timeout message potentially racing with an older one.

This manifested as a bug in the accompanying test, where on the 40th
interleaving Concuerror would ask a 'wrong' timer-simulating process to send the
timer expired message.

Fixes #42.