Accurate clock value for dependencies that are collected
When rescheduling, each reversible race is a dependent pair of events (one
'early' and one 'late'). If the 'late' may also race with even earlier events,
the 'early' should be considered an event that happens-before the 'late'.

Should not make a difference, unless we try to reorder wakeup sequence events.