Report error if an instrumented module contains NIFs
Stateless systematic testing requires that any state of the program that has
already been visited can be reached again just by making the same scheduling
choices. In order for this to work, one must control all other sources of
nondeterminism.

NIFs unfortunately cannot, in general, be guaranteed to return the same result
for each call with the same arguments. With this patch Concuerror aborts
attempts to run code that is loading NIFs.

Fixes #77.