Implement a simple bounding technique
This technique, named 'simple', works in the following way: after exploring an
initial interleaving, races are identified and more interleavings planned by
'diverging' from this initial interleaving. Each such new interleaving diverges
in one step and defines the first few steps of the new exploration. The 'simple'
bound limits how many such divergences are permitted.

This is not equivalent to 'preemption' bounding, as no effort is made to reduce
or reason about preemptions. It is also not equivalent to delay bounding, as
more interleavings can diverge at the same point, and each will only consume
'one' unit of bound budget.

This method works with 'optimal' DPOR, as in this case it is the wakeup sequence
that is considered the 'initial' interleaving (costing 1 unit of budget at the
divergence point).