Add 'enabledness check' for BPOR conservative additions
Due to how happens-before is calculated, it could be the case that
BPOR tried to reschedule a process at a point where it was not
enabled. It happened like this:

- There is a race (which establishes a happens-before order)
- An earlier operation from the same block is enabling the 'late
  operation', but
- This is not detected anywhere, due to how happens-before is assigned
  (from the last operation backwards).

The clocks were not enough to detect that it was impossible to avoid
the preemption in that case, since they only contained the race as the
latest op, and were not aware of the previous ones.