Add concuerror_options(_forced)
This patch is a rewriting of most of the logic in concuerror_options.
The goal was to simplify some of the rules and improve the relevant
messages.

It replaces the concuerror_options() function with two attributes,
one that can be overwritten by the command line and one that cannot.

The tests for options have also been extended.

Fixes #98.
Fixes #12.