Warn when instrumenting modules with calls to erlang:load_nif/2
With this patch Concuerror will emits a warning when
instrumenting modules that may try to load NIFs.

NIFs cannot, in general, be guaranteed to return the same result
for each call with the same arguments, nor can they be
detected/intercepted. Stateless model checking requires that any
state of the program can reliably be reached just by making the
same scheduling choices. In order for this to work, one must
control all other sources of nondeterminism.

Fixes #77.