Refine dependencies of ets operations using a match spec
An ets operation using a match spec (such as match, select or
select_delete) depends only with ets operations modifying matching
tuples, or with ets operations where the tuples are not known.

It would be possible to further improve the dependency tests between a
match spec using ets operation and an operation where only the key is
known by reducing the match spec to the key field only. But this is
not yet implemented.