Pre-instrument error_handler
This is in order to avoid replay errors for code passing through the
code module.

A call to an undefined function will lead to a call to the
error_handler module, which, among other responsibilities is
responsible for automatic code loading, calling
`code:ensure_loaded/1`.  If the error_handler module has not been
instrumented, such a call will 'opaquely' result in an exception being
raised, as modules are always loaded (& instrumented) before
Concuerror-controlled processes reach them.  However, if the 'code'
module is instrumented, one can see some side-effecting operations
inside code:ensure_loaded/1 (the first of which is a monitor placed on
the code_server process).

The bug here is the following:

- in the first scheduling an undefined function raises an undef
  exception

- ALSO in the first scheduling, the code module is instrumented, after
  this exception has been raised

- when replaying the code that leads to the undef exception, the
  monitor event will be triggered instead, causing a 'replay' error.

The solution is to ensure that error_handler is always instrumented,
so that if it triggered it will instrument 'code' and see any
side-effecting events.